backend/requirements.txt
PyJWT 
cryptography 
requests
Flask
pymysql
flask-cors
python-dotenv